Move the CSS to an external stylesheet.
`<style>` elements currently "leak" out of their `<svg>` parents and
apply to the whole document. Scoped stylesheets would fix this, but
they’re not there yet.

Having the stylesheet in its own file rather than in the middle of
JavaScript code makes it easier to tweak.

Also, repeating the same stylesheet over and over is silly.