Add Stack operator for vertical concatenation
The Stack operator provides vertical concatenation where every item
starts with a new line.

A Stack must be the last child when used within a Sequence, Diagram or another Stack.
A Error is thrown if this restriction is violated

Every nodes has an additional height value that represents the offset
on the Y axis where the new line should be connected.

The offsetX value indicates where to start the line on the X axis.
This is needed because after the Stack the x value could be smaller.
All operators but the new Stack operator have a offsetX of 0.

The Stack operator is not available in the Python Port